feat: Notification backend finished

+++ backend/src/main/java/edu/ntnu/idatt2106_2023_06/backend/service/notification/NotificationService.java
@@ -1,8 +1,13 @@
 package edu.ntnu.idatt2106_2023_06.backend.service.notification;
 
+import edu.ntnu.idatt2106_2023_06.backend.dto.items.ItemRemoveDTO;
+import edu.ntnu.idatt2106_2023_06.backend.dto.notification.NotificationDTO;
 import edu.ntnu.idatt2106_2023_06.backend.dto.notification.UpdateNotificationDTO;
 import edu.ntnu.idatt2106_2023_06.backend.exception.UnauthorizedException;
+import edu.ntnu.idatt2106_2023_06.backend.exception.not_found.FridgeItemsNotFoundException;
+import edu.ntnu.idatt2106_2023_06.backend.exception.not_found.ItemNotFoundException;
 import edu.ntnu.idatt2106_2023_06.backend.exception.not_found.NotificationNotFoundException;
+import edu.ntnu.idatt2106_2023_06.backend.mapper.NotificationMapper;
 import edu.ntnu.idatt2106_2023_06.backend.model.fridge.FridgeItems;
 import edu.ntnu.idatt2106_2023_06.backend.model.notification.Notification;
 import edu.ntnu.idatt2106_2023_06.backend.model.users.User;
@@ -16,6 +21,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.Comparator;
 import java.util.List;
 
 @Service
@@ -40,16 +46,17 @@ public class NotificationService implements INotificationService {
         checkNotification(fridgeId);
         User user = userInfoService.getAuthenticatedUserObject();
         List<FridgeItems> items = fridgeItemsRepository.findAllByFridge_FridgeId(fridgeId).orElseThrow(
-                () -> new IllegalArgumentException("No items in fridge")
+                () -> new FridgeItemsNotFoundException(fridgeId)
         );
         for (FridgeItems item : items) {
-            if(item.getExpirationDate().isBefore(LocalDateTime.now().plusDays(1))
+            if(item.getExpirationDate().isBefore(LocalDateTime.now().plusDays(2))
                     && notificationRepository.findByUserAndFridgeItem_Item_ItemId(
                             user, item.getItem().getItemId()).isEmpty()) {
                 Notification notification = Notification.builder()
                         .fridgeItem(item)
                         .user(user)
                         .isRead(false)
+                        .createdDate(LocalDateTime.now())
                         .build();
                 notificationRepository.save(notification);
             }
@@ -59,11 +66,22 @@ public class NotificationService implements INotificationService {
     /**
      * Get all notifications of a user. Uses the JWT token to get the user.
      */
-    public List<Notification> getAllNotifications() {
+    public List<Notification> getAllNotificationsOfSingeUser() {
         if(!userInfoService.isAuthenticated())
             throw new UnauthorizedException(userInfoService.getAuthenticatedUserUsername());
         User user = userInfoService.getAuthenticatedUserObject();
         return notificationRepository.findByUser(user);
+    }
+
+    /**
+     * Get all notifications of a user as DTOs. Uses the JWT token to get the user.
+     *
+     * @return A list of notifications as DTOs.
+     */
+    public List<NotificationDTO> getAllNotificationsOfSingeUserAsDTO() {
+        List<Notification> notifications = getAllNotificationsOfSingeUser();
+        notifications.sort(Comparator.comparing(Notification::getCreatedDate));
+        return NotificationMapper.toNotificationDTO(notifications);
     }
 
     /**
@@ -84,7 +102,7 @@ public class NotificationService implements INotificationService {
      * Sets all notifications for a user as read.
      */
     public void setAllNotificationsAsRead() {
-        List<Notification> notifications = getAllNotifications();
+        List<Notification> notifications = getAllNotificationsOfSingeUser();
         for (Notification notification : notifications) {
             notification.setIsRead(true);
             notificationRepository.save(notification);
@@ -95,11 +113,16 @@ public class NotificationService implements INotificationService {
      * Deletes a notification for every user in a fridge.
      * This method should be called when removing an item from a fridge.
      *
-     * @param notification The notification to delete.
+     * @param itemRemoveDTO The item to delete.
      */
-    public void deleteNotificationForEveryUserInFridge(Notification notification) {
-        checkNotification(notification.getFridgeItem().getFridge().getFridgeId());
-        List<Notification> notifications = notificationRepository.findByFridgeItem(notification.getFridgeItem());
+    public void deleteNotificationForEveryUserInFridge(ItemRemoveDTO itemRemoveDTO) {
+        checkNotification(itemRemoveDTO.fridgeId());
+        FridgeItems fridgeItem = fridgeItemsRepository.findByItem_ProductName(itemRemoveDTO.itemName()).orElseThrow(
+                () -> new FridgeItemsNotFoundException(itemRemoveDTO.itemName())
+        );
+        if (itemRemoveDTO.quantity() != fridgeItem.getQuantity())
+            return;
+        List<Notification> notifications = notificationRepository.findByFridgeItem(fridgeItem);
         notificationRepository.deleteAll(notifications);
     }
 
@@ -113,6 +136,21 @@ public class NotificationService implements INotificationService {
         User user = userInfoService.getAuthenticatedUserObject();
         List<Notification> notifications = notificationRepository.findByUser(user);
         notificationRepository.deleteAll(notifications);
+    }
+
+    /**
+     * Deletes a notification. for a user.
+     *
+     * @param notificationId The id of the notification to delete.
+     */
+    public void deleteNotification(Long notificationId) {
+        if(!userInfoService.isAuthenticated())
+            throw new UnauthorizedException(userInfoService.getAuthenticatedUserUsername());
+        Notification notification = notificationRepository.findById(notificationId).orElseThrow(
+                () -> new NotificationNotFoundException(notificationId)
+        );
+        checkNotification(notification.getFridgeItem().getFridge().getFridgeId());
+        notificationRepository.delete(notification);
     }
 
     /**
